# docs/openapi.generated.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "fieldPath": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "stage": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "kind",
          "detail",
          "retryable",
          "requestId"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "operatorToken": {
        "in": "header",
        "name": "x-pcrstudio-operator-token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Transport-only contract. Scientific payload schemas are versioned separately.",
    "title": "PCRStudio HTTP API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/engines": {
      "get": {
        "operationId": "engines",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Machine-readable PCRStudio error"
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/enzymes": {
      "post": {
        "operationId": "enzymes",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Machine-readable PCRStudio error"
          }
        },
        "tags": [
          "bench"
        ]
      }
    },
...
    "/api/v1/modifiers": {
      "get": {
        "operationId": "modifiers",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Machine-readable PCRStudio error"
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/modules": {
      "get": {
        "operationId": "modules",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Machine-readable PCRStudio error"
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
...
    "/api/v1/presets": {
      "get": {
        "operationId": "presets",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Machine-readable PCRStudio error"
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
...
  }
}